Worked on solver
Z3Solver implements the ISolver interface.  It uses the Z3ExprVisitor
class which uses visitor pattern to collect data out of the expression
passed to it.

The Z3ExprVisitor should be able to dynamically construct a complete Z3
expression to be solved.  Solving it piece by piece isn't what we want.

In Z3ExprVisitor I have added an example of checking for an Expression
of type "OR", it is an incomplete implementation to get you started.